Feat: 회원가입 기능 구현 완료 - domain 폴더 삭제함
