+++ src/main/java/com/codeit/mission/deokhugam/user/dto/UserDto.java
@@ -1,8 +1,8 @@
-
-package com.codeit.mission.deokhugam.domain.user.dto;
+package com.codeit.mission.deokhugam.user.dto;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record UserDto(UUID id, String email, String nickname, LocalDateTime createdAt) {
+
 }
