+++ src/main/java/com/codeit/mission/deokhugam/user/dto/UserUpdateRequest.java
@@ -1,5 +1,4 @@
-
-package com.codeit.mission.deokhugam.domain.user.dto;
+package com.codeit.mission.deokhugam.user.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -9,4 +8,5 @@ public record UserUpdateRequest(
     @Size(min = 2, max = 20, message = "닉네임은 2자 이상 20자 이하로 입력해주세요.")
     String nickname
 ) {
+
 }
