+++ src/main/java/com/codeit/mission/deokhugam/user/enums/UserStatus.java
@@ -1,0 +1,13 @@
+package com.codeit.mission.deokhugam.user.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum UserStatus {
+  ACTIVE("활성"),
+  DELETED("탈퇴");
+
+  private final String description;
+}
