+++ src/main/java/com/codeit/mission/deokhugam/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.codeit.mission.deokhugam.user.repository;
+
+import com.codeit.mission.deokhugam.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  boolean existsByEmail(String email);
+
+  boolean existsByNickname(String nickname);
+}
